Extract EvaluateLogBuilder into a separate module
